--- apps/client/src/entities/signin/index.ts
@@ -1,1 +1,0 @@
-export * from './model/types';
