--- apps/client/src/entities/signin/model/types.ts
@@ -1,7 +1,0 @@
-export interface OAuthCodeRequestInput {
-  email: string;
-  password: string;
-  clientId: string;
-  redirectUrl: string;
-  isInternal: boolean;
-}
