+++ apps/client/src/middleware.ts
@@ -7,19 +7,22 @@ export async function middleware(request: NextRequest) {
   const { pathname } = request.nextUrl;
   const accessToken = request.cookies.get(COOKIE_KEYS.ACCESS_TOKEN)?.value;
 
-  const publicPaths = ['/signup', '/signin/reset-password', '/signup/success'];
+  // 구버전 OAuth 경로 처리 (임시 리다이렉트)
+  if (pathname.startsWith('/oauth')) {
+    const targetUrl = new URL('http://localhost:3004/oauth/authorize');
+    request.nextUrl.searchParams.forEach((value, key) => {
+      targetUrl.searchParams.set(key, value);
+    });
+    return NextResponse.redirect(targetUrl);
+  }
 
   // OAuth callback 감지 (code 파라미터가 있으면 외부 OAuth callback)
   if (request.nextUrl.searchParams.has('code')) {
     return NextResponse.next();
   }
 
-  // OAuth 플로우는 미들웨어 적용 안 함
-  if (pathname.startsWith('/oauth') || pathname.startsWith('/api/')) {
-    return NextResponse.next();
-  }
-
-  if (publicPaths.includes(pathname)) {
+  // API 경로는 미들웨어 적용 안 함
+  if (pathname.startsWith('/api/')) {
     return NextResponse.next();
   }
 
@@ -67,6 +70,7 @@ export async function middleware(request: NextRequest) {
 
 export const config = {
   matcher: [
+    '/oauth/:path*',
     '/((?!api|_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp|ico)$).*)',
   ],
 };
