--- apps/client/src/views/oauth/index.ts
@@ -1,1 +1,0 @@
-export { default as OAuthAuthorizePage } from './ui/OAuthAuthorizePage';
