+++ apps/oauth/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "oauth",
+  "version": "1.0.0",
+  "type": "module",
+  "private": true,
+  "scripts": {
+    "dev": "next dev --port 3004 --turbopack",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint --max-warnings 0",
+    "check-types": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@hookform/resolvers": "^5.2.2",
+    "@repo/shared": "workspace:*",
+    "lucide-react": "^0.546.0",
+    "next": "15.5.9",
+    "react": "19.2.1",
+    "react-dom": "19.2.1",
+    "react-hook-form": "^7.67.0",
+    "zod": "^4.1.13"
+  },
+  "devDependencies": {
+    "@next/eslint-plugin-next": "15.5.9",
+    "@repo/eslint-config": "workspace:*",
+    "@repo/tailwind-config": "workspace:*",
+    "@repo/typescript-config": "workspace:*",
+    "@tailwindcss/postcss": "^4.1.5",
+    "@types/node": "^22.15.30",
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^9.33.0",
+    "postcss": "^8.5.3",
+    "tailwindcss": "^4.1.5",
+    "typescript": "5.9.2"
+  }
+}
