+++ apps/oauth/postcss.config.js
@@ -1,0 +1,3 @@
+import { postcssConfig } from '@repo/tailwind-config/postcss';
+
+export default postcssConfig;
