+++ apps/oauth/src/app/oauth/authorize/page.tsx
