+++ apps/oauth/src/entities/signup/model/schema.ts
