+++ apps/oauth/src/shared/styles/globals.css
@@ -1,0 +1,3 @@
+@import 'tailwindcss';
+@source "../../../../../packages/shared/src";
+@import '../../../../../packages/shared/src/styles/globals.css';
