+++ apps/oauth/src/views/oauth/index.ts
@@ -1,0 +1,1 @@
+export * from './ui/OAuthAuthorizePage';
