+++ apps/oauth/src/views/oauth/ui/OAuthAuthorizePage/index.tsx
@@ -4,12 +4,10 @@ import { cn } from '@repo/shared/utils';
 
 import { OAuthAuthorizeForm } from '@/widgets/oauth';
 
-const OAuthAuthorizePage = () => {
+export const OAuthAuthorizePage = () => {
   return (
     <div className={cn('bg-background flex min-h-screen items-center justify-center px-4')}>
       <OAuthAuthorizeForm />
     </div>
   );
 };
-
-export default OAuthAuthorizePage;
