+++ apps/oauth/src/views/reset-password/index.ts
