+++ apps/oauth/src/widgets/oauth/ui/OAuthAuthorizeForm/index.tsx
