+++ packages/shared/src/hooks/index.ts
@@ -11,3 +11,4 @@ export * from './useUpdateApiKey';
 export * from './useDeleteApiKey';
 export * from './useDeleteApiKeyById';
 export * from './useUpdateApiKeyExpirationById';
+export * from './useDebounce';
